Implement walk-forward grid search API
